% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test.R
\name{test_impl}
\alias{test_impl}
\title{Test impl}
\usage{
test_impl(x)
}
\arguments{
\item{x}{A value}
}
\description{
This is the implementation that powers \code{\link[testinherit:test]{testinherit::test()}}.
}
\seealso{
\code{\link[testinherit:test]{testinherit::test()}}
}
